LPS-95416 Extracting common logic
